requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.4
django-cors-headers==2.1.0
django-rest-swagger==2.1.2
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
drf-nested-routers==0.90.0
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
jira==1.0.10
MarkupSafe==1.0
oauthlib==2.0.2
openapi-codec==1.3.2
pbr==3.1.1
PyJWT==1.5.2
pytz==2017.2
requests==2.18.3
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
simplejson==3.11.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.22
